Grid: Default the `dtype` from the `grid` for `Dense/TimeData` objects

+++ examples/seismic/acoustic/operators.py
@@ -72,7 +72,7 @@ def ForwardOperator(model, source, receiver, time_order=2, space_order=4,
     m, damp = model.m, model.damp
 
     # Create symbols for forward wavefield, source and receivers
-    u = TimeData(name='u', grid=model.grid, dtype=model.dtype,
+    u = TimeData(name='u', grid=model.grid,
                  save=save, time_dim=source.nt if save else None,
                  time_order=2, space_order=space_order)
     src = PointSource(name='src', ntime=source.nt, ndim=source.ndim,
@@ -114,7 +114,7 @@ def AdjointOperator(model, source, receiver, time_order=2, space_order=4, **kwar
     m, damp = model.m, model.damp
 
     v = TimeData(name='v', grid=model.grid, save=False,
-                 time_order=2, space_order=space_order, dtype=model.dtype)
+                 time_order=2, space_order=space_order)
     srca = PointSource(name='srca', ntime=source.nt, ndim=source.ndim,
                        npoint=source.npoint)
     rec = Receiver(name='rec', ntime=receiver.nt, ndim=receiver.ndim,
@@ -153,13 +153,11 @@ def GradientOperator(model, source, receiver, time_order=2, space_order=4, **kwa
     m, damp = model.m, model.damp
 
     # Gradient symbol and wavefield symbols
-    grad = DenseData(name='grad', grid=model.grid, dtype=model.dtype)
-    u = TimeData(name='u', grid=model.grid, save=True,
-                 time_dim=source.nt, time_order=2,
-                 space_order=space_order, dtype=model.dtype)
+    grad = DenseData(name='grad', grid=model.grid)
+    u = TimeData(name='u', grid=model.grid, save=True, time_dim=source.nt,
+                 time_order=2, space_order=space_order)
     v = TimeData(name='v', grid=model.grid, save=False,
-                 time_order=2, space_order=space_order,
-                 dtype=model.dtype)
+                 time_order=2, space_order=space_order)
     rec = Receiver(name='rec', ntime=receiver.nt, ndim=receiver.ndim,
                    npoint=receiver.npoint)
 
@@ -207,10 +205,10 @@ def BornOperator(model, source, receiver, time_order=2, space_order=4, **kwargs)
 
     # Create wavefields and a dm field
     u = TimeData(name="u", grid=model.grid, save=False,
-                 time_order=2, space_order=space_order, dtype=model.dtype)
+                 time_order=2, space_order=space_order)
     U = TimeData(name="U", grid=model.grid, save=False,
-                 time_order=2, space_order=space_order, dtype=model.dtype)
-    dm = DenseData(name="dm", grid=model.grid, dtype=model.dtype)
+                 time_order=2, space_order=space_order)
+    dm = DenseData(name="dm", grid=model.grid)
 
     s = t.spacing
     # Get computational time-step value
